spotify-flask-model/requirements.txt
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.4
pandas==0.25.3
requests==2.22.0
scikit-learn==0.22.1
scipy==1.4.1
matplotlib==3.1.2
Flask-Cors==3.0.8